backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
openai
scikit-learn
hdbscan
numpy
# supabase
python-dotenv
